config: arquitetura remodulada
