+++ arch/src/controllers/routes/apiRoutes.ts
@@ -1,5 +1,6 @@
 import { Router, Request, Response } from 'express';
-import { GetFollowersData } from '../../request/followers';
+import { GetFollowersData } from '../../requests/FollowersRequest';
+import { getUserData } from '../../requests/UserRequest';
 
 const routers = Router();
 
@@ -8,14 +9,16 @@ routers.get('/', (req: Request, res: Response) => {
 });
 
 routers.get('/followers', async (req: Request, res: Response) => {
-    const username = req.query.username as string;
+    const name = req.query.name as string;
     const page = Number(req.query.page) || 1;
 
-    res.json(await GetFollowersData(username, page));
+    res.json(await GetFollowersData(name, page));
 });
 
-routers.get('/user', (req: Request, res: Response) => {
-    res.json({'message': 'Hello World!'});
+routers.get('/user', async(req: Request, res: Response) => {
+    const name = req.query.name as string;
+    
+    res.json(await getUserData(name));
 });
 
 export { routers };
